Make JET an empty module on prereleased Julia versions
unless JET_DEV_MODE set
